backend/requirements.txt
fastapi
uvicorn[standard]
google-api-python-client
earthengine-api
python-dotenv
